Before initial synchronizing flow.xml.gz, check for valid local flow.xml.gz and rename it to allow the synchronization treat the node as new without a flow. Also, if synchronization failed, do the same to for later retrying of synchronization that pulls down the cluster's flow.

+++ nifi-nar-bundles/nifi-framework-bundle/nifi-framework/nifi-framework-core/src/main/java/org/apache/nifi/persistence/StandardXMLFlowConfigurationDAO.java
@@ -16,6 +16,8 @@
  */
 package org.apache.nifi.persistence;
 
+import java.io.ByteArrayInputStream;
+import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
@@ -23,6 +25,8 @@ import java.io.OutputStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.StandardOpenOption;
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
 import java.util.zip.GZIPInputStream;
 import java.util.zip.GZIPOutputStream;
 
@@ -85,7 +89,33 @@ public final class StandardXMLFlowConfigurationDAO implements FlowConfigurationD
             throws IOException, FlowSerializationException, FlowSynchronizationException, UninheritableFlowException, MissingBundleException {
 
         final FlowSynchronizer flowSynchronizer = new StandardFlowSynchronizer(encryptor, nifiProperties, extensionManager);
-        controller.synchronize(flowSynchronizer, dataFlow);
+        
+        // Used for formatting current date as part of backed up flow.xml.gz's file name
+        DateTimeFormatter dateFormatter = DateTimeFormatter.ofPattern("yyyyMMddHHmmss");
+
+        // Make sure local flow XML is valid as it'd be loaded for initial cluster synchronization.
+        // If it's invalid, rename it to something else to allow cluster synchronization to proceed
+        // anyway and NiFi to come up with empty flow instead of dying out due to IOException
+        if (!controller.isFlowSynchronized() && !isValidFlowXml()) {
+            moveFlowXml(dateFormatter.format(LocalDateTime.now()) + ".malformed.gz", 
+                        "being malformed XML");
+        }
+
+        try {
+            controller.synchronize(flowSynchronizer, dataFlow);
+        } catch (UninheritableFlowException e) {
+            // For this error, the node can't be synchronized because its flow.xml.gz is in
+            // conflict with cluster flow. Instead of requiring manual removal of the file and 
+            // restarting NiFi, we just move the file out of the way as if the node has a blank
+            // flow to allow it to use the cluster flow.
+            boolean moved = moveFlowXml(dateFormatter.format(LocalDateTime.now()) + ".uninherited.gz", 
+                                        "cluster flow is uninheritable by local flow");
+            if (!moved) {
+                LOG.error("Failed to rename uninherited flow.xml.gz. Please remove or move it manually " +
+                          "and restart NiFi for this node to be synchronized with the cluster.");
+                return;
+            }
+        }
 
         if (StandardFlowSynchronizer.isEmpty(dataFlow)) {
             // If the dataflow is empty, we want to save it. We do this because when we start up a brand new cluster with no
@@ -107,7 +137,18 @@ public final class StandardXMLFlowConfigurationDAO implements FlowConfigurationD
 
         try (final InputStream inStream = Files.newInputStream(flowXmlPath, StandardOpenOption.READ);
                 final InputStream gzipIn = new GZIPInputStream(inStream)) {
-            FileUtils.copy(gzipIn, os);
+            // Make sure flow XML is well-formed before writing it out
+            final ByteArrayOutputStream baos = new ByteArrayOutputStream();
+            FileUtils.copy(gzipIn, baos);
+            if (isValidXml(baos.toByteArray())) {
+                ByteArrayInputStream bais = new ByteArrayInputStream(baos.toByteArray());
+                FileUtils.copy(bais, os);
+            }
+        } catch (IOException e) {
+            // Just ignore the corrupt file. Cluster/FlowController synchronization will
+            // overwrite it when time comes
+            LOG.warn(flowXmlPath.getFileName() + 
+                    " is corrupt or has malformed XML. Ignored loading: " + e.toString());
         }
     }
 
@@ -131,10 +172,7 @@ public final class StandardXMLFlowConfigurationDAO implements FlowConfigurationD
     @Override
     public void save(final FlowController flow) throws IOException {
         LOG.trace("Saving flow to disk");
-        try (final OutputStream outStream = Files.newOutputStream(flowXmlPath, StandardOpenOption.WRITE, StandardOpenOption.CREATE);
-                final OutputStream gzipOut = new GZIPOutputStream(outStream)) {
-            save(flow, gzipOut);
-        }
+        save(flow, false);
         LOG.debug("Finished saving flow to disk");
     }
 
@@ -154,26 +192,34 @@ public final class StandardXMLFlowConfigurationDAO implements FlowConfigurationD
             throw new NullPointerException();
         }
 
-        Path tempFile;
-        Path configFile;
-
-        configFile = flowXmlPath;
-        tempFile = configFile.getParent().resolve(configFile.toFile().getName() + ".new.xml.gz");
-
+        Path configFile = flowXmlPath;
+        Path tempFile = configFile.getParent().resolve(configFile.toFile().getName() + ".new.xml.gz");
+        
         try (final OutputStream fileOut = Files.newOutputStream(tempFile);
                 final OutputStream outStream = new GZIPOutputStream(fileOut)) {
-
             final StandardFlowSerializer xmlTransformer = new StandardFlowSerializer(encryptor);
             controller.serialize(xmlTransformer, outStream);
-
-            Files.deleteIfExists(configFile);
-            FileUtils.renameFile(tempFile.toFile(), configFile.toFile(), 5, true);
         } catch (final FlowSerializationException fse) {
             throw new IOException(fse);
+        }
+        
+        // Validate the written temp file to be valid XML before updating the live file
+        try (final InputStream inStream = Files.newInputStream(tempFile);
+                final InputStream gzipIn = new GZIPInputStream(inStream)) {
+            ByteArrayOutputStream baos = new ByteArrayOutputStream();
+            FileUtils.copy(gzipIn, baos);
+
+            if (isValidXml(baos.toByteArray())) {
+                Files.deleteIfExists(configFile);
+                FileUtils.renameFile(tempFile.toFile(), configFile.toFile(), 5, true);
+            } else {
+                throw new FlowSerializationException("Saving failed for " + 
+                        configFile.toFile().getName() + ": Invalid XML was written.");
+            }
         } finally {
             Files.deleteIfExists(tempFile);
         }
-
+        
         if (archive) {
             try {
                 archiveManager.archive();
@@ -185,5 +231,44 @@ public final class StandardXMLFlowConfigurationDAO implements FlowConfigurationD
             }
         }
     }
+    
+    /**
+     * Checks if the local flow is a valid XML.
+     * @return
+     */
+    private boolean isValidFlowXml() {
+        boolean valid = false;
+        try (final InputStream inStream = Files.newInputStream(flowXmlPath, StandardOpenOption.READ);
+                final InputStream gzipIn = new GZIPInputStream(inStream);
+                final ByteArrayOutputStream baos = new ByteArrayOutputStream()) {
+            FileUtils.copy(gzipIn, baos);
+            valid = isValidXml(baos.toByteArray());
+        } catch (IOException e) {
+            LOG.warn(flowXmlPath.getFileName() + 
+                    " is corrupt or has malformed XML: " + e.toString());
+        }
+        return valid;
+    }
 
+    /**
+     * Moves or renames the local flow.xml.gz to the same file name 
+     * but with the specified extension.
+     * @param movedToFileExt
+     * @param reasonMessage
+     * @return
+     */
+    private boolean moveFlowXml(String movedToFileExt, String reasonMessage) {
+        String movedToFlowXmlName = flowXmlPath.toFile().getName() + movedToFileExt;
+        Path movedToFlowXmlPath = flowXmlPath.getParent().resolve(movedToFlowXmlName);
+        try {
+            FileUtils.renameFile(flowXmlPath.toFile(), movedToFlowXmlPath.toFile(), 3);
+            LOG.warn("Moved " + flowXmlPath.toFile().getName() + " to " + movedToFlowXmlName + 
+                     " for " + reasonMessage + ".");
+        } catch (IOException e) {
+            LOG.warn("Unable to move " + flowXmlPath.toFile().getName() + 
+                     " for " + reasonMessage + ": " + e.toString() + ".");
+            return false;
+        }
+        return true;
+    }
 }
